iss1288 - Tidy code
